+++ docs/Changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-* What's new in release 6.7
+* What's new in release 6.7 (19 May 2019)
 
   [general]
   - #180: include tests in source package